pyproject.toml
[project]
name = "cambfurs-catbot"
version = "0.1.0"
requires-python = ">= 3.11"
readme = "README.md"
dependencies = [
    "python-telegram-bot>=21.2",
]
